[Bugfix] Fix pr comment
