+++ be/src/testutil/exprs_test_helper.h
@@ -21,7 +21,6 @@
 #include "common/object_pool.h"
 #include "common/status.h"
 #include "exprs/array_expr.h"
-#include "exprs/jit/jit_expr.h"
 #include "exprs/map_expr.h"
 #include "exprs/mock_vectorized_expr.h"
 #include "gen_cpp/Descriptors_types.h"
@@ -171,71 +170,7 @@ public:
         }
         return expr;
     }
-
-    static void verify_with_jit(ColumnPtr ptr, Expr* expr, RuntimeState* runtime_state,
-                                const std::function<void(ColumnPtr const&)>& test_func, bool need_jit = true) {
-        // Verify the original result.
-        test_func(ptr);
-
-        if (!need_jit) {
-            return;
-        }
-        auto jit_engine = JITEngine::get_instance();
-        if (!jit_engine->support_jit()) {
-            return;
-        }
-
-        ObjectPool pool;
-        auto* jit_expr = JITExpr::create(&pool, expr);
-
-        ExprContext exprContext(jit_expr);
-        std::vector<ExprContext*> expr_ctxs = {&exprContext};
-
-        ASSERT_OK(Expr::prepare(expr_ctxs, runtime_state));
-        ASSERT_OK(Expr::open(expr_ctxs, runtime_state));
-        ASSERT_TRUE(jit_expr->is_jit_compiled());
-
-        Chunk chunk;
-        chunk.append_column(ptr, 0);
-        ptr = jit_expr->evaluate(&exprContext, &chunk);
-        // Verify the result after JIT.
-        test_func(ptr);
-
-        Expr::close(expr_ctxs, runtime_state);
-    }
-
-    static void verify_result_with_jit(const ColumnPtr& ptr, Expr* expr, RuntimeState* runtime_state) {
-        auto jit_engine = JITEngine::get_instance();
-        if (!jit_engine->support_jit()) {
-            return;
-        }
-
-        ObjectPool pool;
-        auto* jit_expr = JITExpr::create(&pool, expr);
-        ExprContext exprContext(jit_expr);
-        std::vector<ExprContext*> expr_ctxs = {&exprContext};
-
-        ASSERT_OK(Expr::prepare(expr_ctxs, runtime_state));
-        ASSERT_OK(Expr::open(expr_ctxs, runtime_state));
-        ASSERT_TRUE(jit_expr->is_jit_compiled());
-
-        Chunk chunk;
-        chunk.append_column(ptr, 0);
-        auto jit_ptr = jit_expr->evaluate(&exprContext, &chunk);
-
-        ASSERT_TRUE(jit_ptr->is_constant() == ptr->is_constant());
-
-        // ASSERT_TRUE(jit_ptr->is_nullable() == ptr->is_nullable());
-        ASSERT_TRUE(jit_ptr->size() == ptr->size());
-        for (int i = 0; i < jit_ptr->size(); ++i) {
-            ASSERT_TRUE(jit_ptr->is_null(i) == ptr->is_null(i));
-            ASSERT_TRUE(jit_ptr->equals(i, *ptr, i));
-        }
-
-        Expr::close(expr_ctxs, runtime_state);
-    }
 };
-
 class TExprBuilder {
 public:
     TExprBuilder& operator<<(const LogicalType& slot_type) {
