+++ be/test/exprs/arithmetic_expr_test.cpp
@@ -20,9 +20,9 @@
 #include "butil/time.h"
 #include "column/column_hash.h"
 #include "column/fixed_length_column.h"
+#include "exprs/exprs_test_helper.h"
 #include "exprs/mock_vectorized_expr.h"
 #include "runtime/runtime_state.h"
-#include "testutil/exprs_test_helper.h"
 
 namespace starrocks {
